reorganise projects
module products and add transactions
