+++ src/app/dashboard/dashboard.component.html
@@ -68,7 +68,7 @@
         </a>
       </div>
       <div class="col-md-3">
-        <a [routerLink]="['/dashboard-stocks']">
+        <a [routerLink]="['/inventories']">
           <div class="card card-body card-widget bg-brown card-inverse">
             <div class="mr-auto">
               <h4 class="fw-300 mb-0">Inventory</h4>
