+++ src/app/inventories/products/new-phones/new-phones.component.html
