+++ src/app/inventories/products/phones/phones.component.scss
