+++ src/app/inventories/products/products.component.ts
