+++ src/app/inventories/products/products.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { ProductsRoutes} from './products.routing';
-import {PortableServices} from '../services/portable.services';
+import {PortableServices} from '../../services/portable.services';
 import {ProductsComponent} from './products.component';
 import {PhonesComponent} from './phones/phones.component.';
 import {NewPhonesComponent} from './new-phones/new-phones.component';
