+++ src/app/inventories/products/products.routing.ts
