+++ src/assets/i18n/en.json
@@ -76,7 +76,7 @@
   "PRODUCTS": "Products",
     "PHONES": "Phones",
 
-  "INVENTORIES": "Inventories",
+  "TRANSACTIONS": "Transactions",
 
   "RECIPIENTS": "Providers / Customers",
