+++ src/assets/i18n/fr.json
@@ -76,7 +76,7 @@
   "PRODUCTS": "Produits",
     "PHONES": "Téléphones",
 
-  "INVENTORIES": "Inventaires",
+  "TRANSACTIONS": "Transactions",
 
   "RECIPIENTS": "Founisseurs/Clients",
